api/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
psycopg2-binary==2.9.10

# Authentication
fastapi-users[sqlalchemy]==13.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data processing
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0

# HTTP client (for TMDB API)
httpx==0.28.1

# Environment variables
python-dotenv==1.0.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator>=1.1.0,<2.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
